booking-api/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.10
flask-cors==4.0.0